Revert "(MAINT) Move version.rb to work with ci"
This reverts commit 5be440a433141e7dab5534d2309282d0865adca7.
