+++ lib/hocon/version.rb
